Change the file structure
Make it more clear for working
