+++ .gitignore
@@ -482,4 +482,4 @@ $RECYCLE.BIN/
 
 .idea/
 
-owl-eye-docker-lambda-v3/hello_world/app/opt/
+owl-eye-multiple-models-version/hello_world/app/opt/
