--- owl-eye-docker-lambda-v3/.aws-sam/build.toml
@@ -1,13 +1,0 @@
-# This file is auto generated by SAM CLI build command
-
-[function_build_definitions]
-[function_build_definitions.6be413aa-71c8-4e18-9e8e-05646b120b43]
-packagetype = "Image"
-functions = ["OwlEyeFunctionv3"]
-
-[function_build_definitions.6be413aa-71c8-4e18-9e8e-05646b120b43.metadata]
-Dockerfile = "Dockerfile"
-DockerContext = "/Users/chenbeining/Downloads/GitHub/LearnAWS/owl-eye-docker-lambda-v3/hello_world"
-DockerTag = "python3.7-v1"
-
-[layer_build_definitions]
