+++ owl-eye-multiple-models-version/hello_world/app/requirements.txt
@@ -1,0 +1,5 @@
+-f https://download.pytorch.org/whl/torch_stable.html
+torch==1.8.0+cpu
+torchvision==0.9.0+cpu
+opencv-python==4.5.1.48
+pillow
