+++ owl-eye-multiple-models-version/template.yaml
@@ -1,0 +1,46 @@
+ AWSTemplateFormatVersion: '2010-09-09'
+Transform: AWS::Serverless-2016-10-31
+Description: >
+  python3.7
+
+  Sample SAM Template for owl-eye-docker-lambda-v3
+
+# More info about Globals: https://github.com/awslabs/serverless-application-model/blob/master/docs/globals.rst
+Globals:
+  Function:
+    Timeout: 50
+    MemorySize: 5000
+  Api:
+    BinaryMediaTypes:
+      - image/png
+      - image/jpg
+      - image/jpeg
+Resources:
+  OwlEyeFunctionv3:
+    Type: AWS::Serverless::Function # More info about Function Resource: https://github.com/awslabs/serverless-application-model/blob/master/versions/2016-10-31.md#awsserverlessfunction
+    Properties:
+      PackageType: Image
+      Events:
+        HelloWorld:
+          Type: Api # More info about API Event Source: https://github.com/awslabs/serverless-application-model/blob/master/versions/2016-10-31.md#api
+          Properties:
+            Path: /owleye
+            Method: post
+    Metadata:
+      Dockerfile: Dockerfile
+      DockerContext: ./hello_world
+      DockerTag: python3.7-v1
+
+Outputs:
+  # ServerlessRestApi is an implicit API created out of Events key under Serverless::Function
+  # Find out more about other implicit resources you can reference within SAM
+  # https://github.com/awslabs/serverless-application-model/blob/master/docs/internals/generated_resources.rst#api
+  OwlEyeApi:
+    Description: "API Gateway endpoint URL for Prod stage for owl eye function"
+    Value: !Sub "https://${ServerlessRestApi}.execute-api.${AWS::Region}.amazonaws.com/Prod/owleye/"
+  OwlEyeFunctionv3:
+    Description: "owl eye Lambda Function ARN"
+    Value: !GetAtt OwlEyeFunctionv3.Arn
+  OwlEyeFunctionv3IamRole:
+    Description: "Implicit IAM Role created for owl eye function"
+    Value: !GetAtt OwlEyeFunctionv3Role.Arn
